+++ mayan/apps/common/tests/literals.py
@@ -4,6 +4,8 @@ import os
 
 from django.conf import settings
 
+EXCLUDE_TEST_TAG = 'exclude'
+
 TEST_ERROR_LOG_ENTRY_RESULT = 'test_error_log_entry_result_text'
 TEST_VIEW_NAME = 'test view name'
 TEST_VIEW_URL = 'test-view-url'