+++ HISTORY.rst
@@ -26,6 +26,7 @@
 * Remove newlines from the rendered templates.
 * Reject emails attachments of size 0. Thanks to Robert Schoeftner
   (@robert.schoeftner)for the report and solution. GitLab issue #574.
+* Add missing document index API view create permission.
 
 3.1.9 (2018-11-01)
 ==================